// employee/src/shared-resources/export-file/common/enum/data-table-name.enum.ts
export enum DataTableNameEnum {
  //Authentication Export Modules
  USER = 'USER',
  ROLE = 'ROLE',
  AUDIT_LOG = 'AUDIT LOG',

  //Employee Export Modules
  EMPLOYEE_MASTER_INFORMATION = 'EMPLOYEE MASTER INFORMATION',
  WORK_SHIFT_TYPE = 'WORK SHIFT TYPE',
  WORKING_SHIFT = 'WORKING SHIFT',
  GEOGRAPHIC = 'GEOGRAPHIC',
  PAYMENT_METHOD = 'PAYMENT METHOD',
  INSURANCE = 'INSURANCE',
  VACCINATION = 'VACCINATION',
  PUBLIC_HOLIDAY = 'PUBLIC HOLIDAY',
  OVERTIME_REQUEST_TYPE = 'OVERTIME REQUEST TYPE',
  LEAVE_REQUEST_TYPE = 'LEAVE REQUEST TYPE',
  BENEFIT_COMPONENT_TYPE = 'BENEFIT COMPONENT TYPE',
  BENEFIT_COMPONENT = 'BENEFIT COMPONENT',
  BENEFIT_INCREMENT_POLICY = 'BENEFIT INCREMENT POLICY',
  REASON_TEMPLATE = 'REASON TEMPLATE',
  APPROVAL_REQUEST_LIST = 'APPROVAL REQUEST LIST',
  SCHEDULER_JOB = 'SCHEDULER JOB',
  IMPORT_EMPLOYEE = 'IMPORT EMPLOYEE',
  POSITION_LEVEL = 'POSITION LEVEL',
  FINGER_PRINT_DEVICE = 'FINGER PRINT DEVICE',
  COMPANY_STRUCTURE_COMPONENT = 'COMPANY STRUCTURE COMPONENT',
  EMPLOYEE_RESIGNATION = 'EMPLOYEE RESIGNATION',
  EMPLOYEE_WARNING = 'EMPLOYEE WARNING',
  EMPLOYEE_MOVEMENT = 'EMPLOYEE MOVEMENT',
  COMPANY_STRUCTURE = 'COMPANY STRUCTURE',

  //Attendance Export Modules
  MISSED_SCAN_REQUEST = 'MISSED SCAN REQUEST',
  OVERTIME_REQUEST = 'OVERTIME REQUEST',

  //Leave Export Modules
  MISSION_REQUEST = 'MISSION REQUEST',
  DAY_OFF_REQUEST = 'DAY OFF REQUEST',
  LEAVE_REQUEST = 'LEAVE REQUEST',

  //Payroll Export Modules
  PAYROLL_BENEFIT = 'PAYROLL BENEFIT',
  PAYROLL_BENEFIT_ADJUSTMENT = 'PAYROLL BENEFIT ADJUSTMENT',
  BENEFIT_ADJUSTMENT_TYPE = 'BENEFIT_ADJUSTMENT_TYPE',
  PAYROLL_REPORT = 'PAYROLL REPORT',
  PAYROLL_CYCLE_CONFIGURATION = 'PAYROLL_CYCLE_CONFIGURATION',
  PAYROLL_DEDUCTION = 'PAYROLL DEDUCTION',
  PAYROLL_DEDUCTION_TYPE = 'PAYROLL DEDUCTION TYPE',
  SALARY_TAX_WITHHELD_RANK = 'SALARY_TAX_WITHHELD_RANK',

  //Attendance Report Export Modules
  ATTENDANCE_REPORT = 'ATTENDANCE REPORT',
  ATTENDANCE_REPORT_DETAIL = 'ATTENDANCE REPORT DETAIL',
  ATTENDANCE_RECORD = 'ATTENDANCE RECORD',

  //Employee Report Export Modules
  EMPLOYEE_BIRTHDAY_REPORT = 'EMPLOYEE_BIRTHDAY_REPORT',
  EMPLOYEE_POST_PROBATION_REPORT = 'EMPLOYEE_POST_PROBATION_REPORT',
  EMPLOYEE_WORK_ANNIVERSARY_REPORT = 'EMPLOYEE_WORK_ANNIVERSARY_REPORT',
  LEAVE_REMAIN_REPORT = 'LEAVE REMAIN REPORT'
}